backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
passlib
python-jose[cryptography]
python-dotenv
httpx
pytest
pytest-asyncio